reorg, remove old music slice files
