+++ app/screens/LibraryArtists.tsx
@@ -2,7 +2,7 @@ import FilterButton, { OptionData } from '@app/components/FilterButton'
 import GradientFlatList from '@app/components/GradientFlatList'
 import ListItem from '@app/components/ListItem'
 import { useFetchList2 } from '@app/hooks/list'
-import { Artist } from '@app/models/music'
+import { Artist } from '@app/models/library'
 import { ArtistFilterType } from '@app/models/settings'
 import { selectSettings } from '@app/state/settings'
 import { useStore, useStoreDeep } from '@app/state/store'
